app/requirements-app.txt
django >=5.1,<6.0
django-leaflet >=0.31,<1.0
djangorestframework >=3.15,<4.0
django-filter >=24.3,<25